Setup an init-build project that can be incremental
currently the init-build script will do the following:
- Build the local tasks
- Write out the targget props file
- GenerateRootFs
- Apply Patches
- Build dummy packages
- Build reference assembly packages

It will do all of that once and afterwards a clean will be needed
in order for it to run again. Individual targets can also be run
on the init-builds.proj file in order to repeat just parts of the
project steps.